Add files via upload
Supermarket is a place where the entire products are available. The aiming to automate its existing manual system by the help of computerized equipment and full-fledged computer software, full-filling their requirements so that their valuable data can be stored for a longer period with easy accessing and manipulation. 
Supermarket management is the system where all the aspects related to the proper management of supermarket is done. These aspects involve managing information about various products, employee, supplier and billing etc.